🏷️ 회원가입 관련 타입 추가
