+++ src/types/user.d.ts
@@ -91,6 +91,133 @@ interface DeleteAccountApiResponse extends BaseApiResponse {
     };
 }
 
+/**
+ * Signup form types
+ */
+interface SignUpInputs {
+    /** User's email address */
+    email: string;
+    /** User's verification code */
+    verificationCode: string;
+    /** User's password */
+    password: string;
+    /** User's confirm password */
+    confirmPassword: string;
+    /** User's nickname */
+    nickname: string;
+    /** User's introduce */
+    introduce: string;
+    /** User's gender */
+    genderBool: boolean;
+    /** User's possibility of caring dogs */
+    possible: boolean;
+    /** User's preference of dog sizes */
+    dogSizes: string[];
+    /** User's mbti */
+    mbti: string;
+}
+
+interface emailValidationType {
+    required: string;
+    pattern: {
+        value: RegExp;
+        message: string;
+    };
+}
+
+interface verificationCodeValidationType {
+    required: string;
+}
+
+interface passwordValidationType {
+    required: string;
+    minLength: {
+        value: number;
+        message: string;
+    };
+    pattern: {
+        value: RegExp;
+        message: string;
+    };
+}
+
+interface confirmPasswordValidationType {
+    required: string;
+    validate: (
+        value: string | boolean | string[],
+    ) => boolean | string | Promise<boolean | string>;
+}
+
+interface nicknameValidationType {
+    required: string;
+    minLength: {
+        value: number;
+        message: string;
+    };
+    maxLength: {
+        value: number;
+        message: string;
+    };
+    pattern: {
+        value: RegExp;
+        message: string;
+    };
+}
+
+interface introduceValidationType {
+    maxLength: {
+        value: number;
+        message: string;
+    };
+}
+
+//	POST Send VerificationCode to Email
+interface EmailVerificationCodeApiRequest {
+    accountId: string;
+}
+interface EmailVerificationCodeApiResponse extends BaseApiResponse {
+    data: {
+        responseCode: string;
+        message: string;
+    };
+}
+
+//	POST Check VerificationCode
+interface CheckVerificationCodeApiRequest {
+    accountId: string;
+    certificationNumber: string;
+}
+interface CheckVerificationCodeApiResponse extends BaseApiResponse {
+    data: {
+        responseCode: string;
+        message: string;
+    };
+}
+
+// //	POST Signup
+interface SignupApiRequest {
+    accountId: string;
+    password: string;
+    certificationNumber: string;
+    nickname: string;
+    gender: string;
+    preferredSizes: string[];
+    introduction: string;
+    isCareAvailable: boolean;
+    mbti: string;
+    latitude: string;
+    longitude: string;
+    profileImg: string;
+}
+interface SignupApiResponse extends BaseApiResponse {
+    data: {
+        responseCode: string;
+        message: string;
+        id?: number;
+        imgURL?: string;
+    };
+}
+
 interface IUser {
     id: number;
     accountId: string;
@@ -184,6 +311,19 @@ export type {
     DeleteAccountInputs,
     DeleteAccountApiRequest,
     DeleteAccountApiResponse,
+    SignUpInputs,
+    emailValidationType,
+    verificationCodeValidationType,
+    passwordValidationType,
+    confirmPasswordValidationType,
+    nicknameValidationType,
+    introduceValidationType,
+    EmailVerificationCodeApiRequest,
+    EmailVerificationCodeApiResponse,
+    CheckVerificationCodeApiRequest,
+    CheckVerificationCodeApiResponse,
+    SignupApiRequest,
+    SignupApiResponse,
     IUser,
     LikedUser,
     ProfileApiQueryString,
